refactor(motion_estimator): de-duplicate motion_estimator and extract kernels (#47)
Eighth de-duplication step of #47 — the biggest structural win.

- NEW src/motion_model/motion_model.ts: motion_model + affine2d +
  homography2d kernel classes, moved verbatim from the monolith (they were
  module-local classes there). Kernels instantiate linalg via direct module
  import instead of the jsfeatNext.linalg static slot.
- src/motion_estimator/motion_estimator.ts: replace the type-only stub with
  the REAL implementation (get_subset, find_inliers, ransac, lmeds); lmeds
  instantiates math via direct module import.
- src/core/core.ts: the temporary any-typed affine2d/homography2d static
  slots (a known TODO since #62) now have precise typeof types via type-only
  imports.
- src/jsfeatNext.ts: down to ~380 lines — pure aggregator except for
  optical_flow_lk, the last inline module.

Verified behavior-preserving: tsc --noEmit clean; npm test 57/57 (3 parity
tests pin ransac/lmeds with identical models AND inlier masks vs original
jsfeat on seeded data); UMD bundle smoke-checked (instanceof for estimator
and both kernels, affine RANSAC recovers the synthetic model).
