+++ src/motion_estimator/motion_estimator.ts
@@ -1,10 +1,25 @@
+import jsfeatNext from "../core/core";
 import { IHomography2d } from "../homography2d/homography2d";
 import { matrix_t } from "../matrix_t/matrix_t";
 import { point_t } from "../point_t/point_t";
 import { ransac_params_t } from "./ransac_params_t";
-export class motion_estimator {
+import { JSFEAT_CONSTANTS } from "../constants/constants";
+import { homography2d } from "../motion_model/motion_model";
+import { math } from "../math/math";
+
+/**
+ * Real implementation, moved out of the src/jsfeatNext.ts monolith (issue #47).
+ * This file previously held a type-only stub — the implementation below is the
+ * inline code from the monolith, verbatim (the only change: lmeds instantiates
+ * the math module directly instead of via the jsfeatNext.math static slot).
+ */
+export class motion_estimator extends jsfeatNext {
+    constructor() {
+        super();
+    }
+
     get_subset(
-        kernel: IHomography2d,
+        kernel: homography2d,
         from: point_t[],
         to: point_t[],
         need_cnt: number,
@@ -12,10 +27,44 @@ export class motion_estimator {
         from_sub: point_t[],
         to_sub: point_t[]
     ): boolean {
-        throw new Error("Method not implemented.");
+        const max_try = 1000;
+        const indices = [];
+        let i = 0,
+            j = 0,
+            ssiter = 0,
+            idx_i = 0,
+            ok = false;
+        for (; ssiter < max_try; ++ssiter) {
+            i = 0;
+            for (; i < need_cnt && ssiter < max_try; ) {
+                ok = false;
+                idx_i = 0;
+                while (!ok) {
+                    ok = true;
+                    idx_i = indices[i] = Math.floor(Math.random() * max_cnt) | 0;
+                    for (j = 0; j < i; ++j) {
+                        if (idx_i == indices[j]) {
+                            ok = false;
+                            break;
+                        }
+                    }
+                }
+                from_sub[i] = from[idx_i];
+                to_sub[i] = to[idx_i];
+                if (!kernel.check_subset(from_sub, to_sub, i + 1)) {
+                    ssiter++;
+                    continue;
+                }
+                ++i;
+            }
+            break;
+        }
+
+        return i == need_cnt && ssiter < max_try;
     }
+
     find_inliers(
-        kernel: IHomography2d,
+        kernel: homography2d,
         model: matrix_t,
         from: point_t[],
         to: point_t[],
@@ -24,8 +73,21 @@ export class motion_estimator {
         err: Int32Array | Float32Array,
         mask: number[]
     ): number {
-        throw new Error("Method not implemented.");
+        let numinliers: number = 0,
+            i = 0,
+            f = 0;
+        const t = thresh * thresh;
+
+        kernel.error(from, to, model, err, count);
+
+        for (; i < count; ++i) {
+            f = <number>(<unknown>(err[i] <= t));
+            mask[i] = f;
+            numinliers += f;
+        }
+        return numinliers;
     }
+
     ransac(
         params: ransac_params_t,
         kernel: any,
@@ -36,8 +98,94 @@ export class motion_estimator {
         mask: matrix_t,
         max_iters: number
     ): boolean {
-        throw new Error("Method not implemented.");
+        if (typeof max_iters === "undefined") {
+            max_iters = 1000;
+        }
+
+        if (count < params.size) return false;
+
+        const model_points = params.size;
+        let niters = max_iters,
+            iter = 0;
+        let result: boolean = false;
+
+        const subset0: any = [];
+        const subset1: any = [];
+        let found = false;
+
+        const mc = model.cols,
+            mr = model.rows;
+        const dt = model.type | JSFEAT_CONSTANTS.C1_t;
+
+        const m_buff = this.cache.get_buffer((mc * mr) << 3);
+        const ms_buff = this.cache.get_buffer(count);
+        const err_buff = this.cache.get_buffer(count << 2);
+        const M = new matrix_t(mc, mr, dt, m_buff.data);
+        const curr_mask = new matrix_t(count, 1, JSFEAT_CONSTANTS.U8C1_t, ms_buff.data);
+
+        let inliers_max = -1,
+            numinliers = 0;
+        let nmodels = 0;
+
+        const err = err_buff.f32;
+
+        // special case
+        if (count == model_points) {
+            if (kernel.run(from, to, M, count) <= 0) {
+                this.cache.put_buffer(m_buff);
+                this.cache.put_buffer(ms_buff);
+                this.cache.put_buffer(err_buff);
+                return false;
+            }
+
+            M.copy_to(model);
+            if (mask) {
+                while (--count >= 0) {
+                    mask.data[count] = 1;
+                }
+            }
+            this.cache.put_buffer(m_buff);
+            this.cache.put_buffer(ms_buff);
+            this.cache.put_buffer(err_buff);
+            return true;
+        }
+
+        for (; iter < niters; ++iter) {
+            // generate subset
+            found = this.get_subset(kernel, from, to, model_points, count, subset0, subset1);
+            if (!found) {
+                if (iter == 0) {
+                    this.cache.put_buffer(m_buff);
+                    this.cache.put_buffer(ms_buff);
+                    this.cache.put_buffer(err_buff);
+                    return false;
+                }
+                break;
+            }
+
+            nmodels = kernel.run(subset0, subset1, M, model_points);
+            if (nmodels <= 0) continue;
+
+            // TODO handle multimodel output
+
+            numinliers = this.find_inliers(kernel, M, from, to, count, params.thresh, err, curr_mask.data);
+
+            if (numinliers > Math.max(inliers_max, model_points - 1)) {
+                M.copy_to(model);
+                inliers_max = numinliers;
+                if (mask) curr_mask.copy_to(mask);
+                niters = params.update_iters((count - numinliers) / count, niters);
+                result = true;
+            }
+        }
+
+        this.cache.put_buffer(m_buff);
+        this.cache.put_buffer(ms_buff);
+        this.cache.put_buffer(err_buff);
+
+        return result;
     }
+
     lmeds(
         params: ransac_params_t,
         kernel: any,
@@ -48,6 +196,106 @@ export class motion_estimator {
         mask: matrix_t,
         max_iters: number
     ): boolean {
-        throw new Error("Method not implemented.");
+        if (typeof max_iters === "undefined") {
+            max_iters = 1000;
+        }
+
+        if (count < params.size) return false;
+
+        const model_points = params.size;
+        let niters = max_iters,
+            iter = 0;
+        let result: boolean = false;
+        const _math = new math();
+
+        const subset0: any = [];
+        const subset1: any = [];
+        let found = false;
+
+        const mc = model.cols,
+            mr = model.rows;
+        const dt = model.type | JSFEAT_CONSTANTS.C1_t;
+
+        const m_buff = this.cache.get_buffer((mc * mr) << 3);
+        const ms_buff = this.cache.get_buffer(count);
+        const err_buff = this.cache.get_buffer(count << 2);
+        const M = new matrix_t(mc, mr, dt, m_buff.data);
+        const curr_mask = new matrix_t(count, 1, JSFEAT_CONSTANTS.U8_t | JSFEAT_CONSTANTS.C1_t, ms_buff.data);
+
+        let numinliers = 0;
+        let nmodels = 0;
+
+        const err = err_buff.f32;
+        let min_median = 1000000000.0,
+            sigma = 0.0,
+            median = 0.0;
+
+        params.eps = 0.45;
+        niters = params.update_iters(params.eps, niters);
+
+        // special case
+        if (count == model_points) {
+            if (kernel.run(from, to, M, count) <= 0) {
+                this.cache.put_buffer(m_buff);
+                this.cache.put_buffer(ms_buff);
+                this.cache.put_buffer(err_buff);
+                return false;
+            }
+
+            M.copy_to(model);
+            if (mask) {
+                while (--count >= 0) {
+                    mask.data[count] = 1;
+                }
+            }
+            this.cache.put_buffer(m_buff);
+            this.cache.put_buffer(ms_buff);
+            this.cache.put_buffer(err_buff);
+            return true;
+        }
+
+        for (; iter < niters; ++iter) {
+            // generate subset
+            found = this.get_subset(kernel, from, to, model_points, count, subset0, subset1);
+            if (!found) {
+                if (iter == 0) {
+                    this.cache.put_buffer(m_buff);
+                    this.cache.put_buffer(ms_buff);
+                    this.cache.put_buffer(err_buff);
+                    return false;
+                }
+                break;
+            }
+
+            nmodels = kernel.run(subset0, subset1, M, model_points);
+            if (nmodels <= 0) continue;
+
+            // TODO handle multimodel output
+
+            kernel.error(from, to, M, err, count);
+            median = _math.median(err, 0, count - 1);
+
+            if (median < min_median) {
+                min_median = median;
+                M.copy_to(model);
+                result = true;
+            }
+        }
+
+        if (result) {
+            sigma = 2.5 * 1.4826 * (1 + 5.0 / (count - model_points)) * Math.sqrt(min_median);
+            sigma = Math.max(sigma, 0.001);
+
+            numinliers = this.find_inliers(kernel, model, from, to, count, sigma, err, curr_mask.data);
+            if (mask) curr_mask.copy_to(mask);
+
+            result = numinliers >= model_points;
+        }
+
+        this.cache.put_buffer(m_buff);
+        this.cache.put_buffer(ms_buff);
+        this.cache.put_buffer(err_buff);
+
+        return result;
     }
 }
